@@ -1,4 +1,4 @@
-// swift-tools-version:5.3
+// swift-tools-version: 5.9
 // The swift-tools-version declares the minimum version of Swift required to build this package.
 
 import PackageDescription
@@ -6,24 +6,33 @@ import PackageDescription
 let package = Package(
 	name: "Geodesic",
 	products: [
-		// Products define the executables and libraries a package produces, and make them visible to other packages.
+		// Products define the executables and libraries a package produces, making them visible to other packages.
 		.library(
 			name: "Geodesic",
 			targets: ["Geodesic"]),
 	],
 	dependencies: [
-		// Dependencies declare other packages that this package depends on.
-		// .package(url: /* package url */, from: "1.0.0"),
-		.package(url: "https://github.com/sbooth/CGeodesic", from: "2.0.0")
+		.package(url: "https://github.com/sbooth/CXXGeographicLib", from: "2.4.0")
 	],
 	targets: [
-		// Targets are the basic building blocks of a package. A target can define a module or a test suite.
-		// Targets can depend on other targets in this package, and on products in packages this package depends on.
+		// Targets are the basic building blocks of a package, defining a module or a test suite.
+		// Targets can depend on other targets in this package and products from dependencies.
 		.target(
 			name: "Geodesic",
-			dependencies: ["CGeodesic"]),
+			dependencies: [
+				"CXXGeographicLib",
+			],
+			swiftSettings: [
+				.interoperabilityMode(.Cxx),
+			]),
 		.testTarget(
 			name: "GeodesicTests",
-			dependencies: ["Geodesic"]),
-	]
+			dependencies: [
+				"Geodesic",
+			],
+			swiftSettings: [
+				.interoperabilityMode(.Cxx),
+			]),
+	],
+	cxxLanguageStandard: .cxx17
 )
